[extension] update api
